fix: Fix Chat last message timestamp
The method for updating the last message details wasn't properly
setting the value for the lastMessageDate.

+++ app/src/main/java/project/sherpa/ui/adapters/ChatAdapter.java
@@ -85,7 +85,7 @@ public class ChatAdapter extends RecyclerView.Adapter<ChatAdapter.ChatViewHolder
 
                 oldChat.setActiveMembers(chat.getActiveMembers());
                 oldChat.setMemberCode(chat.getMemberCode());
-                oldChat.setLastMessageDate((long) chat.getLastMessageDate());
+                oldChat.setLastMessageDate(chat.getLastMessageDate());
                 oldChat.setLastMessage(chat.getLastMessage());
                 oldChat.setMessageCount(chat.getMessageCount());
                 oldChat.setLastAuthorId(chat.getLastAuthorId());
